Extract different types of indices in vindex
Extracts the scalar indices and applies them first in `vindex`. Also
drops the scalar indices from the indices both as they are handled and
as the array's rank is reduced.  This is important as these shrink the
shape of the array. Thus it is easier to reason about later steps if we
know this has already happened. Further these immediately result in Dask
Array chunks being dropped from consideration as only the chunk with the
index is kept. This is a nice property to have before other
subselections are applied.

Similarly collect all partial `slice`s and apply them subsequently. Take
into account the array's rank is reduced from the scalar indices so far
collected. In the general indices collect `slice(None)` for each `slice`
partial or otherwise found in the selection. While partial slices were
already handled similar before, stubbing them out with `slice(None)` is
nice for later steps.