Use concatenate in minimum, mean, and maximum
Instead of using Dask Array's `block`, which has some overhead due to
additional checks, simply `concatenate` each dimension after finding how
to pad the ends. This is a bit faster and works nicely in this case.
Plus this rewrite cleans up the code a bit making it more readable in
the process.