Implement a `SynchronousExecutor`
Provides a simple synchronous `Executor` for use with
`concurrent.futures`. Simply executes any function once submitted and
stores the result or exception in a `Future` returned to the user. This
matches closely with our current implementation of synchronous
execution. Also it is simpler to implement than say delaying execution
until the `Future`'s `.result(...)` is queried. Besides as we already
collect and deduplicate all work to do in the local scheduler before
submitting it, there is no material difference between computing
immediately or when the result is requested.