Pass `Future` directly to callback
As any `Exception` (or subclass thereof) that is raised by the callback
will be logged instead of allowed to propagate through, we don't get the
excepted exception handling behavior with `concurrent.futures` when
grabbing the result here. So pass the `Future` instead so that this is
delayed until the `Future`'s result is later retrieved. Should allow the
exception to be handled in the proper context.