Use synchronous scheduler in order test
As this test was using the threaded scheduler previously, it could be
sensitive to the behavior of the scheduler itself, which may result in
some tasks finishing first (even if they were submitted/started in the
correct order). So switch to the synchronous scheduler, which will run
tasks in the order given. This should make this test behave more
consistently.